# spec/openapi.yaml
openapi: 3.1.3
info:
  title: MedVault API
  description: >
    REST API for MedVault — a multi-tenant healthcare platform.
    This contract is the single source of truth for all API endpoints.
  version: 0.1.0
  contact:
    name: MedVault Team
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Register a new user (no tenant assignment)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Auth]
      operationId: authenticateUser
      summary: Authenticate user and return available tenants
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /auth/select-tenant:
    post:
      tags: [Auth]
      operationId: selectTenant
      summary: Select tenant and receive JWT with tenant + role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectTenantRequest'
      responses:
        '200':
          description: Tenant selected, JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User does not belong to this tenant
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logoutUser
      summary: Invalidate refresh token and log out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
  /users/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      summary: Get current user profile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tenants:
    post:
      tags: [Tenants]
      operationId: createTenant
      summary: Create a new tenant (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Tenant name already exists
  /tenants/{tenant_id}/members:
    get:
      tags: [Tenant Members]
      operationId: listTenantMembers
      summary: List all members of a tenant
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of tenant members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TenantMember'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User is not a member of this tenant

    post:
      tags: [Tenant Members]
      operationId: addTenantMember
      summary: Add a user to a tenant (admin only)
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTenantMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TenantMember'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Principal is not an administrator
  /tenants/{tenant_id}/members/{user_id}:
    delete:
      tags: [Tenant Members]
      operationId: removeTenantMember
      summary: Remove a user from a tenant (admin only)
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Principal is not an administrator
  /tenants/{tenant_id}/reactivate:
    post:
      tags: [Tenant Members]
      operationId: reactivateTenant
      summary: Reactivate a suspended tenant (admin only)
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant reactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      status:
                        type: string
                      created_at:
                        type: string
                        format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Principal is not an administrator
        '404':
          $ref: '#/components/responses/NotFound'
  /tenants/{tenant_id}/suspend:
    post:
      tags: [Tenants]
      operationId: suspendTenant
      summary: Suspend an active tenant (admin only)
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      status:
                        type: string
                      created_at:
                        type: string
                        format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Principal is not an administrator
        '404':
          $ref: '#/components/responses/NotFound'
  /cases:
    get:
      tags: [Cases]
      operationId: listCases
      summary: List cases (filtered by role)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [open, assigned, diagnosed, closed]
      responses:
        '200':
          description: List of cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaseSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Cases]
      operationId: createCase
      summary: Create a new medical case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          description: Case created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cases/{id}:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    get:
      tags: [Cases]
      operationId: getCase
      summary: Get case details
      responses:
        '200':
          description: Case details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not authorized to view this case
        '404':
          $ref: '#/components/responses/NotFound'
  /cases/{id}/symptoms:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      tags: [Symptoms]
      operationId: addSymptom
      summary: Add symptom to case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSymptomRequest'
      responses:
        '201':
          description: Symptom added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymptomResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Not authorized to add symptoms to this case
        '404':
          $ref: '#/components/responses/NotFound'
  /cases/{id}/assign:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      tags: [Cases]
      operationId: assignDoctor
      summary: Assign doctor to case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignDoctorRequest'
      responses:
        '200':
          description: Doctor assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Only administrators can assign doctors
        '404':
          $ref: '#/components/responses/NotFound'
  /cases/{id}/diagnosis:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      tags: [Cases]
      operationId: writeDiagnosis
      summary: Write diagnosis for case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WriteDiagnosisRequest'
      responses:
        '201':
          description: Diagnosis written
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosisResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Only the assigned doctor can write a diagnosis
        '404':
          $ref: '#/components/responses/NotFound'
  /cases/{id}/close:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      tags: [Cases]
      operationId: closeCase
      summary: Close a diagnosed case
      responses:
        '200':
          description: Case closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          description: Case must be diagnosed before closing
        '403':
          description: Only administrators can close cases
        '404':
          $ref: '#/components/responses/NotFound'
  /cases/{id}/images/upload-url:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      tags: [Images]
      operationId: requestUploadURL
      summary: Request pre-signed upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadURLRequest'
      responses:
        '200':
          description: Pre-signed URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadURLResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Not authorized to upload to this case
        '404':
          $ref: '#/components/responses/NotFound'
  /cases/{id}/images:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      tags: [Images]
      operationId: confirmUpload
      summary: Confirm image upload and record metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmUploadRequest'
      responses:
        '201':
          description: Image recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [Images]
      operationId: listImages
      summary: List images for case
      responses:
        '200':
          description: List of images
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /images/{id}/download-url:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Images]
      operationId: getDownloadURL
      summary: Get pre-signed download URL
      responses:
        '200':
          description: Pre-signed download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadURLResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /images/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      tags: [Images]
      operationId: deleteImage
      summary: Delete an image (admin only)
      responses:
        '200':
          description: Image deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only administrators can delete images
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
              description: Temporary JWT (no tenant context)
            tenants:
              type: array
              items:
                type: object
                properties:
                  tenant_id:
                    type: string
                    format: uuid
                  tenant_name:
                    type: string
                  role:
                    type: string
                    enum: [patient, doctor, administrator]
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    UserResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            tenant_id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            role:
              type: string
              enum: [patient, doctor, administrator]
            status:
              type: string
              enum: [active, inactive]
            created_at:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...